+++ locales/en/iffcc.js
@@ -1,0 +1,98 @@
+var iffcc =
+    {
+    iffcc :
+        {
+        text   :
+            {
+            profile : ["name",
+                       "wingspan",
+                       "length",
+                       "speed"]
+            },
+        switch :
+            {
+            selected :
+                {
+                left  : "active",
+                right : "inactive"
+                }
+            },
+        data :
+            {
+            profiles : [{
+                        name        : "A-10",
+                        wingspan    : 57,
+                        length      : 54,
+                        speed       : 310,
+                        selected    : true
+                        },
+                        {
+                        name        : "F-15",
+                        wingspan    : 43,
+                        length      : 64,
+                        speed       : 370,
+                        selected    : true
+                        },
+                        {
+                        name        : "F-16",
+                        wingspan    : 33,
+                        length      : 47,
+                        speed       : 400,
+                        selected    : false
+                        },
+                        {
+                        name        : "F-18",
+                        wingspan    : 37,
+                        length      : 56,
+                        speed       : 320,
+                        selected    : false
+                        },
+                        {
+                        name        : "MIG-29",
+                        wingspan    : 37,
+                        length      : 57,
+                        speed       : 320,
+                        selected    : false
+                        },
+                        {
+                        name        : "SU-27",
+                        wingspan    : 48,
+                        length      : 72,
+                        speed       : 330,
+                        selected    : false
+                        },
+                        {
+                        name        : "SU-25",
+                        wingspan    : 46,
+                        length      : 44,
+                        speed       : 390,
+                        selected    : true
+                        },
+                        {
+                        name        : "AH-64",
+                        wingspan    : 49,
+                        length      : 49,
+                        speed       : 150,
+                        selected    : true
+                        },
+                        {
+                        name        : "UH-60",
+                        wingspan    : 53,
+                        length      : 49,
+                        speed       : 150,
+                        selected    : false
+                        },
+                        {
+                        name        : "MI-8 HIP",
+                        wingspan    : 70,
+                        length      : 60,
+                        speed       : 140,
+                        selected    : false
+                        }]
+            }
+        }
+    }
+
+// // //
+
+module.exports = iffcc