+++ public/javascripts/common/header.js
@@ -55,6 +55,7 @@ Header.prototype.selectTab = function(tab)
             break
         case 4:
             $("tab:nth-of-type(4)").addClass("selected")
+            Iffcc.slideIn()
             break
         }